Update: recognition_service.py, .gitignore. Add: create_database.py,  support poetry
