+++ lib/face_sdk/core/model_handler/face_alignment/FaceAlignModelHandler.py
@@ -1,5 +1,5 @@
 import logging.config
-logging.config.fileConfig("config/logging.conf")
+# logging.config.fileConfig("config/logging.conf")
 logger = logging.getLogger('sdk')
 
 import cv2
