+++ lib/face_sdk/core/model_loader/BaseModelLoader.py
@@ -5,7 +5,7 @@ from pathlib import Path
 sys.path.append('models/network_def')
 import logging.config
 
-logging.config.fileConfig("config/logging.conf")
+# logging.config.fileConfig("config/logging.conf")
 logger = logging.getLogger('sdk')
 from abc import ABCMeta, abstractmethod
